move table types to a separate file
